Improved broadcaster object
git-svn-id: https://svn.code.sf.net/p/beebeep/code@1082 127b15f6-8070-407b-945c-913c06b37495

+++ src/core/NetworkAddress.cpp
@@ -34,7 +34,7 @@ NetworkAddress::NetworkAddress( const NetworkAddress& na )
   (void)operator=( na );
 }
 
-NetworkAddress::NetworkAddress( const QHostAddress& ha, int hp )
+NetworkAddress::NetworkAddress( const QHostAddress& ha, quint16 hp )
  : m_hostAddress( ha ), m_hostPort( hp ), m_info( "" )
 {
 }
